refactor for mutation
